requirements.txt
flask>=3.0.0
openpyxl>=3.1.5
google-genai>=1.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
openai>=1.0.0
langgraph>=0.2.0
